Defer numba guvectorize compilation for cohort functions
